add gcs bucket base
